Order total should always be equal to the cost of the selected item.
Shipping and tax costs should not be included. In the previous
calculation the cost of the previously selected item was subtracted from
the current order total and the new selected item's cost was added.
Because shipping and tax costs can be different per product, the new
order total could include shipping and tax that do not apply to the
new item.